accelerometer/Accelerometer.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Accelerometer Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/accelerometer/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/generic-sensor/generic-sensor-tests.js"></script>
<div id="log"></div>
<iframe src="support-iframe.html" id="frame" style="display:none" sandbox="allow-scripts">
</iframe>
<script>

async_test(t => {
  let sensor = new Accelerometer({includeGravity: false});
  sensor.onactivate = t.step_func_done(() => {
    assert_equals(sensor.includesGravity, false);
    sensor.stop();
    assert_equals(sensor.state, 'idle');
  });
  sensor.onerror = t.step_func_done(unreached);
  sensor.start();
}, "sensor can be successfully created with sensor specific configuration parameter");

runSensorFrequency(Accelerometer);
runGenericSensorTests(Accelerometer);

</script>
